hal: refactor project structure
Structure project using pkgs and command line utilities.
