+++ go.mod
@@ -1,0 +1,19 @@
+module github.com/drivenets/vmw_tsf
+
+go 1.16
+
+require (
+	github.com/Juniper/go-netconf v0.1.1
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cenkalti/backoff/v4 v4.1.1 // indirect
+	github.com/cespare/xxhash/v2 v2.1.1 // indirect
+	github.com/cloudflare/goflow/v3 v3.4.2
+	github.com/openconfig/gnmi v0.0.0-20210527163611-d3a3e30199da
+	github.com/openconfig/ygot v0.10.15
+	github.com/sirupsen/logrus v1.8.1
+	github.com/ziutek/telnet v0.0.0-20180329124119-c3b780dc415b // indirect
+	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a
+	golang.org/x/net v0.0.0-20210614182718-04defd469f4e
+	google.golang.org/grpc v1.38.0
+	google.golang.org/protobuf v1.26.0 // indirect
+)
