+++ pkg/hal/gnmi.go
