+++ pkg/hal/hal_test.go
