+++ pkg/hal/mock.go
