+++ pkg/hal/netconf_api.go
