+++ pkg/hal/structs.go
