--- tsf-client/go.mod
@@ -1,40 +1,0 @@
-module github.com/drivenets/vmw_tsf/tsf-client
-
-go 1.16
-
-require (
-	github.com/Juniper/go-netconf v0.1.1 // indirect
-	github.com/beorn7/perks v1.0.1 // indirect
-	github.com/cenkalti/backoff/v4 v4.1.1 // indirect
-	github.com/cespare/xxhash/v2 v2.1.1 // indirect
-	github.com/cloudflare/goflow/v3 v3.4.2 // indirect
-	github.com/drivenets/vmw_tsf/tsf-hal v0.0.0-20210616105720-98499ee2be99
-	github.com/drivenets/vmw_tsf/tsf-twamp v0.0.0-20210616105720-98499ee2be99 // indirect
-	github.com/golang/glog v0.0.0-20210429001901-424d2337a529 // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/google/go-cmp v0.5.6 // indirect
-	github.com/kylelemons/godebug v1.1.0 // indirect
-	github.com/libp2p/go-reuseport v0.0.2 // indirect
-	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
-	github.com/openconfig/gnmi v0.0.0-20210527163611-d3a3e30199da // indirect
-	github.com/openconfig/goyang v0.2.4 // indirect
-	github.com/openconfig/ygot v0.10.15 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/prometheus/client_golang v1.11.0 // indirect
-	github.com/prometheus/client_model v0.2.0 // indirect
-	github.com/prometheus/common v0.29.0 // indirect
-	github.com/prometheus/procfs v0.6.0 // indirect
-	github.com/sirupsen/logrus v1.8.1 // indirect
-	github.com/ziutek/telnet v0.0.0-20180329124119-c3b780dc415b // indirect
-	golang.org/x/crypto v0.0.0-20210513164829-c07d793c2f9a // indirect
-	golang.org/x/net v0.0.0-20210614182718-04defd469f4e // indirect
-	golang.org/x/sys v0.0.0-20210616094352-59db8d763f22 // indirect
-	golang.org/x/text v0.3.6 // indirect
-	google.golang.org/genproto v0.0.0-20210614182748-5b3b54cad159 // indirect
-	google.golang.org/grpc v1.38.0 // indirect
-	google.golang.org/protobuf v1.26.0 // indirect
-)
-
-replace github.com/drivenets/vmw_tsf/tsf-hal => ../tsf-hal
-
-replace github.com/drivenets/vmw_tsf/tsf-twamp => ../tsf-twamp
