--- tsf-hal/Makefile
@@ -1,4 +1,0 @@
-.PHONY: test
-
-test:
-	go test
