--- tsf-halo-stub/.gitignore
@@ -1,1 +1,0 @@
-go.sum
