--- tsf-twamp/go.mod
@@ -1,5 +1,0 @@
-module github.com/drivenets/vmw_tsf/tsf-twamp
-
-go 1.16
-
-require golang.org/x/net v0.0.0-20210525063256-abc453219eb5
